+++ src/onegov.activity/onegov/activity/upgrade.py
@@ -3,10 +3,10 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 
-from onegov.activity import Booking
+from onegov.activity import Booking, Period
 from onegov.core.orm.types import UUID
 from onegov.core.upgrade import upgrade_task
-from sqlalchemy import Column, Enum, ForeignKey
+from sqlalchemy import Boolean, Column, Enum, ForeignKey
 from sqlalchemy.orm import joinedload
 
 
@@ -59,3 +59,16 @@ def change_booking_states(context):
         ALTER TABLE bookings ALTER COLUMN state
         TYPE booking_state USING state::text::booking_state;
     """)
+
+
+@upgrade_task('Add confirmed flag to period')
+def add_confirmed_flag_to_period(context):
+    context.operations.add_column('periods', Column(
+        'confirmed', Boolean, nullable=True, default=False
+    ))
+
+    for period in context.session.query(Period):
+        period.confirmed = False
+
+    context.session.flush()
+    context.operations.alter_column('periods', 'confirmed', nullable=False)